Switch Backup to use the new Stream framework.
